refactor(rfc3195): extract shared test helpers to testutil_test.go
Move beepFrame and seqFrame helpers from raw_test.go to a shared
testutil_test.go file. Both RAW and COOKED test files use these
helpers.
