+++ rfc3195/raw_test.go
@@ -1,7 +1,6 @@
 package rfc3195
 
 import (
-	"fmt"
 	"strconv"
 	"strings"
 	"testing"
@@ -10,22 +9,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-// beepFrame builds a raw BEEP data frame string.
-// Size is computed automatically from len(payload).
-// For ANS frames, pass ansno >= 0. For other types, ansno is ignored.
-func beepFrame(keyword string, channel, msgno int, more byte, seqno int, ansno int, payload string) string {
-	size := len(payload)
-	header := fmt.Sprintf("%s %d %d %c %d %d", keyword, channel, msgno, more, seqno, size)
-	if keyword == "ANS" {
-		header += " " + strconv.Itoa(ansno)
-	}
-	return header + "\r\n" + payload + "END\r\n"
-}
-
-func seqFrame(channel, ackno, window int) string {
-	return fmt.Sprintf("SEQ %d %d %d\r\n", channel, ackno, window)
-}
 
 // A minimal valid RFC 5424 message
 const validRFC5424 = "<1>1 - - - - - -"
